Display actual puzzle list from CSV
