+++ mobile/src/main/java/ch/epfl/sdp/mobile/application/chess/ChessFacade.kt
@@ -1,17 +1,27 @@
 package ch.epfl.sdp.mobile.application.chess
 
+import android.content.Context
 import ch.epfl.sdp.mobile.application.ChessDocument
 import ch.epfl.sdp.mobile.application.Profile
 import ch.epfl.sdp.mobile.application.ProfileDocument
 import ch.epfl.sdp.mobile.application.authentication.AuthenticatedUser
 import ch.epfl.sdp.mobile.application.authentication.NotAuthenticatedUser
+import ch.epfl.sdp.mobile.application.chess.engine.Color
 import ch.epfl.sdp.mobile.application.chess.engine.Game
+import ch.epfl.sdp.mobile.application.chess.engine.implementation.buildBoard
 import ch.epfl.sdp.mobile.application.chess.notation.AlgebraicNotation.toAlgebraicNotation
+import ch.epfl.sdp.mobile.application.chess.notation.FenNotation
+import ch.epfl.sdp.mobile.application.chess.notation.UCINotation
 import ch.epfl.sdp.mobile.application.chess.notation.mapToGame
 import ch.epfl.sdp.mobile.application.toProfile
 import ch.epfl.sdp.mobile.infrastructure.persistence.auth.Auth
 import ch.epfl.sdp.mobile.infrastructure.persistence.store.*
+import ch.epfl.sdp.mobile.ui.puzzles.Puzzle
+import ch.epfl.sdp.mobile.ui.puzzles.SnapshotPuzzle
+import com.opencsv.CSVReaderHeaderAware
+import io.github.sceneview.utils.useText
 import kotlinx.coroutines.flow.*
+import java.io.FileReader
 
 /**
  * An interface which represents all the endpoints and available features for online chess
@@ -21,6 +31,10 @@ import kotlinx.coroutines.flow.*
  * @param store the [Store] which is used to manage documents.
  */
 class ChessFacade(private val auth: Auth, private val store: Store) {
+
+  /**
+   * Chess matches side of chess facade
+   **/
 
   /**
    * Creates a "local" [Match] for the [AuthenticatedUser] and stores it in the [Store]
@@ -84,6 +98,65 @@ class ChessFacade(private val auth: Auth, private val store: Store) {
       it.filterNotNull().mapNotNull(ChessDocument::uid).map { uid -> StoreMatch(uid, store) }
     }
   }
+
+  /**
+   * Puzzle side of chess facade
+   */
+  suspend fun solvePuzzle(puzzle: Puzzle, user: AuthenticatedUser) {
+    user.update { solvedPuzzles(puzzle) }
+  }
+
+  private fun allPuzzles(context: Context): List<Puzzle> {
+    val bufferedReader = context.assets.open("puzzles/puzzles.csv").bufferedReader()
+    val reader = CSVReaderHeaderAware(bufferedReader)
+    val csvMap = mutableListOf<Map<String, String>>()
+    var line = reader.readMap()
+    while (line != null) {
+      csvMap.add(line)
+      line = reader.readMap()
+    }
+
+    val puzzles =
+      csvMap.map {
+        val puzzleId = it["PuzzleId"] ?: "Error"
+        val fen =
+          FenNotation.parseFen(it["FEN"] ?: "")
+            ?: FenNotation.BoardSnapshot(
+              board = buildBoard {},
+              playing = Color.White,
+              castlingRights =
+              FenNotation.CastlingRights(
+                kingSideWhite = false,
+                queenSideWhite = false,
+                kingSideBlack = false,
+                queenSideBlack = false,
+              ),
+              enPassant = null,
+              halfMoveClock = -1,
+              fullMoveClock = -1,
+            )
+        val moves = UCINotation.parseActions(it["Moves"] ?: "") ?: emptyList()
+        val rating = (it["Rating"] ?: "-1").toInt()
+
+        SnapshotPuzzle(
+          uid = puzzleId,
+          boardSnapshot = fen,
+          puzzleMoves = moves,
+          elo = rating,
+        )
+      }
+
+    return puzzles
+  }
+
+  fun solvedPuzzles(profile: Profile, context: Context): List<Puzzle> {
+    return allPuzzles(context).filter { profile.solvedPuzzles.contains(it.uid) }
+  }
+
+  fun unsolvedPuzzles(profile: Profile, context: Context): List<Puzzle> {
+    return allPuzzles(context).filterNot { profile.solvedPuzzles.contains(it.uid) }
+  }
+
 }
 
 private data class StoreMatch(
