+++ mobile/src/main/java/ch/epfl/sdp/mobile/state/StatefulPuzzleSelectionScreen.kt
@@ -1,10 +1,15 @@
 package ch.epfl.sdp.mobile.state
 
+import android.content.Context
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import ch.epfl.sdp.mobile.application.authentication.AuthenticatedUser
+import ch.epfl.sdp.mobile.application.chess.ChessFacade
+import ch.epfl.sdp.mobile.application.chess.engine.Color
 import ch.epfl.sdp.mobile.ui.game.ChessBoardState
+import ch.epfl.sdp.mobile.ui.puzzles.Puzzle
 import ch.epfl.sdp.mobile.ui.puzzles.PuzzleItem
 import ch.epfl.sdp.mobile.ui.puzzles.PuzzleSelectionScreen
 import ch.epfl.sdp.mobile.ui.puzzles.PuzzleSelectionScreenState
@@ -18,14 +23,17 @@ fun StatefulPuzzleSelectionScreen(
     contentPadding: PaddingValues = PaddingValues(),
 ) {
   val chess = LocalChessFacade.current
+  val context = LocalContext.current
   val onGameItemClickAction = rememberUpdatedState(onPuzzleItemClick)
   val scope = rememberCoroutineScope()
   val state =
       remember(user, chess, scope) {
         SnapshotPuzzleSelectionScreen(
-            user = user,
             onPuzzleClickAction = onGameItemClickAction,
+            user = user,
+            facade = chess,
             scope = scope,
+            context = context,
         )
       }
   PuzzleSelectionScreen(
@@ -38,7 +46,9 @@ fun StatefulPuzzleSelectionScreen(
 private class SnapshotPuzzleSelectionScreen(
     onPuzzleClickAction: State<(puzzle: PuzzleItem) -> Unit>,
     private val user: AuthenticatedUser,
+    private val facade: ChessFacade,
     private val scope: CoroutineScope,
+    private val context: Context,
 ) : PuzzleSelectionScreenState {
 
   val onPuzzleClickAction by onPuzzleClickAction
@@ -47,12 +57,7 @@ private class SnapshotPuzzleSelectionScreen(
     onPuzzleClickAction(puzzle)
   }
 
-  override val puzzles =
-      listOf(
-          PuzzleItemAdapter(uid = "000001", playerColor = ChessBoardState.Color.White, elo = 1500),
-          PuzzleItemAdapter(uid = "000002", playerColor = ChessBoardState.Color.Black, elo = 2000),
-          PuzzleItemAdapter(uid = "000003", playerColor = ChessBoardState.Color.White, elo = 2500),
-      )
+  override val puzzles = facade.unsolvedPuzzles(user, context).map { it.toPuzzleItem() }.sortedBy { it.elo }
 }
 
 data class PuzzleItemAdapter(
@@ -60,3 +65,17 @@ data class PuzzleItemAdapter(
     override val playerColor: ChessBoardState.Color,
     override val elo: Int,
 ) : PuzzleItem
+
+private fun Puzzle.toPuzzleItem(): PuzzleItem {
+  val color =
+      when (this.boardSnapshot.playing) {
+        Color.White -> ChessBoardState.Color.White
+        Color.Black -> ChessBoardState.Color.Black
+      }
+
+  return PuzzleItemAdapter(
+      uid = this.uid,
+      playerColor = color,
+      elo = this.elo,
+  )
+}
